Move LazyConditions to arangod/Aql/Parser
they are only used for the parser.
