+++ arangod/Aql/Parser/LazyConditions.cpp
@@ -20,7 +20,7 @@
 ///
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "LazyConditions.h"
+#include "Aql/Parser/LazyConditions.h"
 
 #include "Aql/Ast.h"
 #include "Aql/Scopes.h"
