+++ arangod/Aql/Parser/Parser.h
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "Aql/LazyConditions.h"
+#include "Aql/Parser/LazyConditions.h"
 #include "Basics/ErrorCode.h"
 
 #include <cstddef>
